+++ packages/fhir-validation/src/utilities.ts
@@ -1,0 +1,3 @@
+export function validateIsObject(v: unknown): v is object {
+  return typeof v === "object" && v !== null;
+}
